# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MysticsHR API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Soft-delete (deactivate) an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deactivated
  /employees/{id}/status:
    patch:
      operationId: updateEmployeeStatus
      tags: [employees]
      summary: Update employee status (lifecycle transition)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeStatusBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # ──────────────────────────────────────────────
  # USERS (HRMS accounts linked to employees)
  # ──────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all HRMS users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HrmsUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create an HRMS user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrmsUser"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the current authenticated user's profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrmsUser"
        "404":
          description: User not found / not set up yet
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrmsUser"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (role or status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HrmsUser"

  # ──────────────────────────────────────────────
  # AUDIT LOGS
  # ──────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
          format: date
        gender:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        designationId:
          type: ["integer", "null"]
        designationTitle:
          type: ["string", "null"]
        employmentType:
          type: string
        status:
          type: string
        dateOfJoining:
          type: ["string", "null"]
          format: date
        ctc:
          type: ["number", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - firstName
        - lastName
        - email
        - employmentType
        - status
        - isDeleted
        - createdAt
        - updatedAt
    EmployeeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - data
        - total
        - limit
        - offset
    CreateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        designationId:
          type: ["integer", "null"]
        employmentType:
          type: string
        dateOfJoining:
          type: ["string", "null"]
        ctc:
          type: ["number", "null"]
        managerId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - email
        - employmentType
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        designationId:
          type: ["integer", "null"]
        employmentType:
          type: ["string", "null"]
        dateOfJoining:
          type: ["string", "null"]
        ctc:
          type: ["number", "null"]
        managerId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    UpdateEmployeeStatusBody:
      type: object
      properties:
        status:
          type: string
        reason:
          type: ["string", "null"]
      required:
        - status

    # ──────────────────────────────────────────────
    # USERS
    # ──────────────────────────────────────────────
    HrmsUser:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        employeeId:
          type: ["integer", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - email
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        clerkUserId:
          type: string
        employeeId:
          type: ["integer", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
      required:
        - clerkUserId
        - email
        - name
        - role
    UpdateUserBody:
      type: object
      properties:
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        employeeId:
          type: ["integer", "null"]

    # ──────────────────────────────────────────────
    # AUDIT LOGS
    # ──────────────────────────────────────────────
...